requirements.txt
flask_sqlalchemy
flasgger
prometheus-flask-exporter
psycopg2
PyYAML
pytest
# ...adicione outras dependências aqui, se necessário...
